[cudaextender] C style casts to static cast

+++ cudaextender/src/ungapped_xdrop_kernels.cu
@@ -402,7 +402,7 @@ __global__ void find_high_scoring_segment_pairs(const char* __restrict__ d_targe
         {
             if (lane_id == 0)
             {
-                if (((int32_t)(((float)total_score[warp_id]) * entropy[warp_id])) >= score_threshold)
+                if (static_cast<int32_t>(static_cast<double>(total_score[warp_id]) * entropy[warp_id]) >= score_threshold)
                 {
                     d_scored_segment[hid].seed_pair.target_position_in_read = ref_loc[warp_id] - left_extent[warp_id];
                     d_scored_segment[hid].seed_pair.query_position_in_read  = query_loc[warp_id] - left_extent[warp_id];
